requirements.txt
click
prettytable
PyYAML
rich